Bhoomi/requirements.txt
Django==4.1.3
psycopg2==2.9.5
django-tinymce==3.5.0
django-sendgrid-v5 0.6.83==0.6.83